[vmesh] using less memory

+++ tofu/geom/_sampling_tools.pyx
@@ -1731,7 +1731,7 @@ cdef inline void vmesh_ind_cart_loop(int np,
                                      double[::1] res3d,
                                      double reso_r_z,
                                      double[::1] dRPhirRef,
-                                     long[::1] Ru,
+                                     int[::1] Ru,
                                      double[::1] dRPhir,
                                      int num_threads) nogil:
     cdef int ii
@@ -1770,7 +1770,7 @@ cdef inline void vmesh_ind_polr_loop(int np,
                                      double[::1] res3d,
                                      double reso_r_z,
                                      double[::1] dRPhirRef,
-                                     long[::1] Ru,
+                                     int[::1] Ru,
                                      double[::1] dRPhir,
                                      int num_threads) nogil:
     cdef int ii
